move mock db to dao
